+++ docs/CHANGELOG.rst
@@ -18,6 +18,7 @@ Added
 - Support different dependency kinds between data objects
 - ``--pull`` option to the ``list_docker_images`` management command
 - Test profiling and process tagging
+- Add 'all' and 'any' jinja filters
 
 Changed
 -------